+++ gwcelery/tasks/gcn/__init__.py
@@ -7,118 +7,37 @@ References
 .. [GCN] https://gcn.gsfc.nasa.gov
 """
 import socket
-import struct
-import time
+import subprocess
 
+import billiard
 from celery.utils.log import get_task_logger
-from celery_eternal import EternalTask, EternalProcessTask
-from gcn.voeventclient import _get_now_iso8601, _recv_packet, _send_packet
+from celery_eternal import EternalProcessTask
+from comet.icomet import IHandler
+from comet.utility import xml_document
+from comet.protocol import VOEventSenderFactory
 from gcn import get_notice_type, NoticeType
-import gcn
+from twisted.application import app as twisted_app
+from twisted.internet import reactor
+from twisted.python.runtime import platformType
+from zope.interface import implementer
 
 from ...import app
 from ..core import DispatchHandler
 
+if platformType == 'win32':
+    from twisted.scripts._twistw import ServerOptions, \
+        WindowsApplicationRunner as ApplicationRunner
+else:
+    from twisted.scripts._twistd_unix import ServerOptions, \
+        UnixApplicationRunner as ApplicationRunner
+
 log = get_task_logger(__name__)
-
-
-IAMALIVE = '''<?xml version="1.0" encoding="UTF-8"?>
-<trn:Transport role="iamalive" version="1.0"
-xmlns:trn="http://telescope-networks.org/schema/Transport/v1.1"
-xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-xsi:schemaLocation="http://telescope-networks.org/schema/Transport/v1.1
-http://telescope-networks.org/schema/Transport-v1.1.xsd">
-<Origin>{}</Origin><TimeStamp>{}</TimeStamp>
-</trn:Transport>'''
-
-
-KEEPALIVE_TIME = 60  # Time in seconds between keepalives
-
-
-def _host_port(address):
-    host, port = address.split(':')
-    return host, int(port)
-
-
-@app.task(base=EternalTask, bind=True)
-def broker(self):
-    """Single-client VOEvent broker for sending notices to GCN.
-
-    This is a basic VOEvent broker. It binds to the address
-    :obj:`~gwcelery.conf.gcn_broker_address` and accepts
-    one connection at a time from any host whose address is listed in
-    :obj:`~gwcelery.conf.gcn_broker_accept_addresses`.
-    """
-    fqdn = socket.getfqdn()
-    host, port = _host_port(app.conf['gcn_broker_address'])
-    accept_hosts = [socket.gethostbyname(host) for host in
-                    app.conf['gcn_broker_accept_addresses']]
-
-    with socket.socket() as sock:
-        sock.settimeout(1.0)
-        sock.bind((host, port))
-        log.info('bound to %s', app.conf['gcn_broker_address'])
-        sock.listen(0)
-        while not self.is_aborted():
-            try:
-                conn, (addr, _) = sock.accept()
-            except socket.timeout:
-                continue
-            if addr in accept_hosts:
-                log.info('accepted connection from remote host %s', addr)
-                break
-            else:
-                log.error('denied connection from remote host %s', addr)
-                conn.close()
-        else:  # self.is_aborted()
-            return
-
-    with conn:
-        conn.settimeout(1.0)
-        conn.setsockopt(socket.SOL_SOCKET, socket.SO_LINGER,
-                        struct.pack('ii', 1, 0))
-        conn.setsockopt(socket.SOL_SOCKET, socket.SO_SNDBUF, 1)
-
-        last_sent = time.monotonic()
-
-        while not self.is_aborted():
-            # Get next payload from queue in first-in, first-out fashion
-            payload = self.backend.client.lindex(_queue_name, 0)
-            if payload is not None:
-                log.info('sending payload of %d bytes', len(payload))
-                _send_packet(conn, payload)
-                last_sent = time.monotonic()
-                self.backend.client.lpop(_queue_name)
-            elif time.monotonic() - last_sent > KEEPALIVE_TIME:
-                timestamp = _get_now_iso8601()
-                payload = IAMALIVE.format(fqdn, timestamp).encode('utf-8')
-                log.info('sending keepalive')
-                _send_packet(conn, payload)
-                last_sent = time.monotonic()
-            else:
-                # Read (but ignore) any inbound packet
-                try:
-                    _recv_packet(conn)
-                except socket.timeout:
-                    pass
-
-
-_queue_name = broker.name + '.voevent-queue'
-
-
-@app.task(bind=True, ignore_result=True, shared=False)
-def send(self, payload):
-    """Send a VOEvent to GCN.
-
-    Under the hood, this task just pushes the payload onto a Redis queue,
-    and :func:`~gwcelery.tasks.gcn.broker` sends it."""
-    self.backend.client.rpush(_queue_name, payload)
 
 
 class _VOEventDispatchHandler(DispatchHandler):
 
-    def process_args(self, payload, root):
-        notice_type = get_notice_type(root)
+    def process_args(self, event):
+        notice_type = get_notice_type(event.element)
 
         # Just cast to enum for prettier log messages
         try:
@@ -126,7 +45,7 @@ class _VOEventDispatchHandler(DispatchHandler):
         except ValueError:
             pass
 
-        return notice_type, (payload,), {}
+        return notice_type, (event.raw_bytes,), {}
 
 
 handler = _VOEventDispatchHandler()
@@ -153,10 +72,106 @@ Declare a new handler like this::
 """
 
 
-@app.task(base=EternalProcessTask, shared=False)
-def listen():
-    """Listen to GCN notices forever. GCN notices are dispatched asynchronously
-    to tasks that have been registered with
-    :meth:`gwcelery.tasks.gcn.handler`."""
-    host, port = _host_port(app.conf['gcn_client_address'])
-    gcn.listen(host, port, handler=handler.dispatch)
+@implementer(IHandler)
+class _VOEventHandler:
+    """ Comet VOEvent handler."""
+    name = 'GWCelery VOEvent dispatch handler'
+    __call__ = handler.dispatch
+
+
+def _get_ephemeral_port():
+    """Get an ephemeral (unused, high numbered) port. Note that there is
+    inherently a race condition in using this function because there is no
+    guarantee that the port is still not in use when the caller attempts to use
+    it."""
+    with socket.socket() as s:
+        s.bind(('127.0.0.1', 0))
+        _, port = s.getsockname()
+    return port
+
+
+@app.task(base=EternalProcessTask, bind=True, shared=False)
+def broker(self):
+    """Run an embedded :doc:`Comet VOEvent broker <comet:usage/broker>` to send
+    and recieve GCNs."""
+    # Look up the broker broadcast port from the global configuration object.
+    _, _, broadcast_port = app.conf['gcn_broker_address'].partition(':')
+
+    # Pick an ephemeral port to listen to for VOEvent submission.
+    # Using an ephemeral port allows us to run multiple instances of
+    # GWCelery on the same machine, which may be useful for testing.
+    author_port = str(_get_ephemeral_port())
+    self.backend.client.set(self.name + '.port', author_port)
+
+    # Assemble the command line options for Twisted.
+    cmd = ['--nodaemon', '--pidfile=', 'comet', '--verbose',
+           '--local-ivo', 'ivo://ligo.org/gwcelery',
+           '--remote', app.conf['gcn_client_address'],
+           '--receive', '--receive-port', author_port,
+           '--author-whitelist', '127.0.0.0/8',
+           '--broadcast', '--broadcast-port', broadcast_port,
+           '--broadcast-test-interval', '0',
+           *(_ for name in app.conf['gcn_broker_accept_addresses']
+             for _ in ('--subscriber-whitelist', socket.gethostbyname(name)))]
+
+    # Run Twisted.
+    log.info('twistd %s', ' '.join(cmd))
+    config = ServerOptions()
+    config.parseOptions(cmd)
+    config.subOptions['handlers'].append(_VOEventHandler())
+    runner = ApplicationRunner(config)
+    runner.run()
+    if runner._exitSignal is not None:
+        twisted_app._exitWithSignal(runner._exitSignal)
+
+
+class _OneShotSender(VOEventSenderFactory):
+    # This class is adapted from the comet-sendvo script.
+    # It allows us to do a single-shot VOEvent submission without
+    # knowing the path where comet-sendvo is installed.
+
+    def clientConnectionLost(self, connector, reason):  # noqa: N802
+        reactor.stop()
+
+    def clientConnectionFailed(self, connector, reason):  # noqa: N802
+        log.warn("Connection failed")
+        reactor.stop()
+
+
+def _send(message, port):
+    # This function is adapted from the comet-sendvo script.
+    # It allows us to do a single-shot VOEvent submission without
+    # knowing the path where comet-sendvo is installed.
+
+    event = xml_document(message)
+    factory = _OneShotSender(event)
+    reactor.connectTCP('localhost', port, factory)
+    reactor.run()
+    if not factory.ack:
+        raise ConnectionError('Comet did not accept the event')
+
+
+@app.task(bind=True, ignore_result=True, shared=False)
+def send(self, message):
+    """Send a VOEvent to the local Comet instance for forwarding to GCN.
+
+    Internally, this works similarly to the :doc:`comet-sendvo
+    <comet:usage/publisher>` command line tool. However, it spawns a separate
+    subprocess because it seems like it is impossible with Twisted to block on
+    a single-shot action more than once in the same Python session because of
+    how the global singleton reactor object works."""
+
+    # Look up the ephemeral port number saved in Redis.
+    port = int(broker.backend.client.get(broker.name + '.port'))
+
+    # Start the process, and wait until it exits.
+    process = billiard.Process(target=_send, args=(message, port), daemon=True)
+    process.start()
+    process.join()
+
+    # If the process failed, then raise an exception. The exception will report
+    # that the called process was comet-sendvo, but of course this is a
+    # harmless little lie.
+    if process.exitcode != 0:
+        raise subprocess.CalledProcessError(
+            process.exitcode, ['comet-sendvo', '--port', str(port)])